requirement.txt
boto3==1.34.84
botocore==1.34.84
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
idna==3.6
imageio==2.33.1
importlib-resources==6.1.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.4
matplotlib==3.7.4
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.9.0.80
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
PyWavelets==1.4.1
requests==2.31.0
s3transfer==0.10.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.7.10
tqdm==4.66.1
transforms==0.2.1
triton==2.2.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
utils==1.0.2
zipp==3.17.0
pika